refactor

+++ src/img4d_lib/color_space.d
@@ -6,7 +6,7 @@ import std.stdio,
        std.range;
 
 double[][] to_grayscale(T)(ref T[][][] color){
-    ulong input_len = color[0][0].length; 
+    uint input_len = color[0][0].length; 
     if (input_len != 3 && input_len != 4) throw new Exception("invalid format.");
     if (input_len == 4)
         color.each!((idx,a) => a.each!((edx,b) => color[idx][edx] = b.remove(3)));
